refactor(liblili2json): move challenge tests to a new file
